Upload observation-level settings file
This settings file generates the predicted proportion of utterances in a class session that contain a particular talk feature. It is used in our lab's internal ML pipeline but the settings should be clear to the reader. This script should be run once per feature, changing the desired predicted feature each time.