Upload utterance-level predictions
This settings file generates the binary predictions of whether a particular talk feature is present in a given utterance. It is used in our lab's internal ML pipeline but the settings should be clear to the reader. This script should be run once per feature, changing the desired predicted feature each time.